Use .apply() on SharedPreferences
.commit() blocks until the edit is completed in order to return the
the write status. .apply(), on the other hand, updates the in-memory
structure, and schedules a background write.

If the return value of .commit() is never used, .apply() should be
used instead to not wait on disk I/O.

I agree to license all rights to my contributions in each modified
file exclusively to KryptCo, Inc.